backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
psutil
scikit-learn
redis
psycopg2-binary
pydantic-settings
sqlalchemy
httpx
groq
